# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/example/send": {
            "get": {
                "description": "Example Deposit List Send",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Example"
                ],
                "summary": "Example Send",
                "operationId": "Example Send",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReqDepositList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResExampleSendSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResExampleSendError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ReqDepositList": {
            "type": "object",
            "properties": {
                "accountHolderId": {
                    "type": "string",
                    "example": "8a8e87bf7c7d97b7017c92cbdcf40abe"
                },
                "accountHolderType": {
                    "type": "string",
                    "example": "CLIENT"
                },
                "paginationDetails": {
                    "description": "ID of an item\n\nIn: query",
                    "type": "string",
                    "example": "OFF"
                }
            }
        },
        "models.ResDeposit": {
            "type": "object",
            "properties": {
                "accountHolderKey": {
                    "type": "string"
                },
                "accountHolderType": {
                    "type": "string"
                },
                "accountState": {
                    "type": "string"
                },
                "accountType": {
                    "type": "string"
                },
                "accruedAmounts": {
                    "type": "object",
                    "properties": {
                        "interestAccrued": {
                            "type": "integer"
                        },
                        "negativeInterestAccrued": {
                            "type": "integer"
                        },
                        "overdraftInterestAccrued": {
                            "type": "integer"
                        },
                        "technicalOverdraftInterestAccrued": {
                            "type": "integer"
                        }
                    }
                },
                "approvedDate": {
                    "type": "string"
                },
                "assignedBranchKey": {
                    "type": "string"
                },
                "balances": {
                    "type": "object",
                    "properties": {
                        "availableBalance": {
                            "type": "number"
                        },
                        "blockedBalance": {
                            "type": "number"
                        },
                        "feesDue": {
                            "type": "integer"
                        },
                        "forwardAvailableBalance": {
                            "type": "integer"
                        },
                        "holdBalance": {
                            "type": "number"
                        },
                        "lockedBalance": {
                            "type": "number"
                        },
                        "overdraftAmount": {
                            "type": "integer"
                        },
                        "overdraftInterestDue": {
                            "type": "integer"
                        },
                        "technicalOverdraftAmount": {
                            "type": "integer"
                        },
                        "technicalOverdraftInterestDue": {
                            "type": "integer"
                        },
                        "totalBalance": {
                            "type": "number"
                        }
                    }
                },
                "creationDate": {
                    "type": "string"
                },
                "currencyCode": {
                    "type": "string"
                },
                "encodedKey": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "interestSettings": {
                    "type": "object",
                    "properties": {
                        "interestPaymentSettings": {
                            "type": "object",
                            "properties": {
                                "interestPaymentDates": {
                                    "type": "array",
                                    "items": {}
                                },
                                "interestPaymentPoint": {
                                    "type": "string"
                                }
                            }
                        },
                        "interestRateSettings": {
                            "type": "object",
                            "properties": {
                                "encodedKey": {
                                    "type": "string"
                                },
                                "interestChargeFrequency": {
                                    "type": "string"
                                },
                                "interestChargeFrequencyCount": {
                                    "type": "integer"
                                },
                                "interestRate": {
                                    "type": "number"
                                },
                                "interestRateSource": {
                                    "type": "string"
                                },
                                "interestRateTerms": {
                                    "type": "string"
                                },
                                "interestRateTiers": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                },
                "internalControls": {
                    "type": "object",
                    "properties": {
                        "maxWithdrawalAmount": {
                            "type": "number"
                        }
                    }
                },
                "lastModifiedDate": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "overdraftInterestSettings": {
                    "type": "object"
                },
                "overdraftSettings": {
                    "type": "object",
                    "properties": {
                        "allowOverdraft": {
                            "type": "boolean"
                        },
                        "overdraftLimit": {
                            "type": "integer"
                        }
                    }
                },
                "productTypeKey": {
                    "type": "string"
                },
                "withholdingTaxSourceKey": {
                    "type": "string"
                }
            }
        },
        "models.ResExampleSendError": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/response.MetaErrMock"
                },
                "response_code": {
                    "type": "string",
                    "example": "BC-500"
                },
                "response_message": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
        "models.ResExampleSendSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ResDeposit"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "response_code": {
                    "type": "string",
                    "example": "BC-200"
                },
                "response_message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "debug_param": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string",
                    "example": "97125121-ea32-4ee0-8706-5b7375e83e94"
                }
            }
        },
        "response.MetaErrMock": {
            "type": "object",
            "properties": {
                "debug_param": {
                    "type": "string",
                    "example": "Request Time Out"
                },
                "trace_id": {
                    "type": "string",
                    "example": "97125121-ea32-4ee0-8706-5b7375e83e94"
                }
            }
        }
    }
}